refactor(CollabEditor): get ydoc versions in CollabHistory when modal opened
- Fix when no snapshots user can edit document in CollabHistory
- Add no histories translation
