+++ components/Collaborate/CollabEditor.js
@@ -21,8 +21,6 @@ import useCurrentUser from 'lib/useCurrentUser';
 import PlaceholderPlugin from './Placeholder';
 import getConfig from 'next/config';
 import CollabHistory from './CollabHistory';
-import gql from 'graphql-tag';
-import { useQuery } from '@apollo/react-hooks';
 
 const {
   publicRuntimeConfig: { PUBLIC_COLLAB_SERVER_URL },
@@ -83,17 +81,6 @@ const colors = [
 
 const color = colors[Math.floor(Math.random() * colors.length)];
 
-const YDOC_VERSIONS_QUERY = gql`
-  query GetYdocVersions($id: String!) {
-    GetYdoc(id: $id) {
-      versions {
-        createdAt
-        snapshot
-      }
-    }
-  }
-`;
-
 const Editor = ({ provider, currentUser, className, innerRef, onUnmount }) => {
   useEffect(() => {
     // console.log('editor mount');
@@ -152,14 +139,6 @@ const CollabEditor = ({ article }) => {
   // onTranscribe setup provider for both Editor and CollabHistory to use.
   // And, to avoid duplicated connection, provider will be destroyed(close connection) when Editor unmounted.
   const [provider, setProvider] = useState(null);
-  const { loading, data: getYdocData, refetch } = useQuery(
-    YDOC_VERSIONS_QUERY,
-    {
-      variables: { id: article.id },
-      ssr: false, // No need to fetch on server
-    }
-  );
-  const versionList = getYdocData?.GetYdoc?.versions || [];
 
   const onTranscribe = () => {
     if (!currentUser) {
@@ -208,10 +187,6 @@ const CollabEditor = ({ article }) => {
       // TODO: listen textChanged event?
       article.text = text;
     }
-
-    // refetch to get latest versionList
-    refetch({ id: article.id });
-
     setShowEditor(false);
   };
 
@@ -263,12 +238,8 @@ const CollabEditor = ({ article }) => {
                 {t`Edit`}
               </Button>
             ) : (
-              isSynced &&
-              !loading && (
-                <CollabHistory
-                  ydoc={provider.document}
-                  versionList={versionList}
-                />
+              isSynced && (
+                <CollabHistory ydoc={provider.document} docName={article.id} />
               )
             )}
           </>
